feat: move FormFieldSize to Form.types.ts, and import syntax cleanup
