--- packages/bezier-react/src/components/Forms/FormFieldSize.ts
@@ -1,1 +1,0 @@
-export type FormFieldSize = 'xl' | 'l' | 'm' | 'xs'
