+++ packages/bezier-react/src/components/Forms/Inputs/TextArea/TextArea.test.tsx
@@ -5,7 +5,7 @@ import { act } from '@testing-library/react'
 
 import { render } from '~/src/utils/test'
 
-import { COMMON_IME_CONTROL_KEYS } from '~/src/components/Forms/Inputs/constants/CommonImeControlKeys'
+import { COMMON_IME_CONTROL_KEYS } from '~/src/components/Forms'
 
 import { TextArea } from './TextArea'
 import type { TextAreaProps } from './TextArea.types'
